+++ packages/terra-tabs/src/terra-dev-site/test/tabs/TabPane/CustomDisplayTabPane.test.jsx
@@ -1,20 +1,24 @@
 import React from 'react';
+import classNames from 'classnames/bind';
 import IconSearch from 'terra-icon/lib/icon/IconSearch';
 import TabPane from '../../../../TabPane';
+import styles from './common/TabPane.test.scss';
+
+const cx = classNames.bind(styles);
 
 const CustomDisplayTabPane = () => (
-  <div role="tablist" style={{ display: 'flex' }}>
+  <div role="tablist" className={cx('content-wrapper')}>
     <TabPane
       label="Custom Display"
       id="customDisplay"
       customDisplay={(
         <div id="customDisplayContent">
-          <span style={{ marginRight: '15px' }}>
+          <span className={cx('custom-display-content')}>
             Custom Display
           </span>
           <IconSearch />
         </div>
-)}
+      )}
     />
   </div>
 );